refactor(tui): colocate single-screen tests into screens/*.rs
Moves the 132 of 333 tests in tui/tests.rs that exercise exactly one
screen (plus their screen-exclusive helpers) into a #[cfg(test)] mod
tests block in that screen's own screens/*.rs file. Shared fixtures
used across screens (state_with_gists, set_pending, pins_mut, etc.)
stay in tests.rs as pub(super), visible to the moved test modules.
Cross-cutting/pure-logic tests (ranking, filtering, scroll, footer)
and the 20 multi-screen navigation tests stay in tests.rs unchanged.

No behavior change — same 333 tests, same assertions, just relocated
to sit next to the screen behaviour they verify (closes #318).
