+++ src/tui/screens/detail.rs
@@ -970,3 +970,614 @@ pub(crate) fn detail_palette_items(state: &AppState) -> Vec<crate::tui::palette:
         key_item("?", "Help", KeyCode::Char('?'), true),
     ]
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::tui::*;
+
+    use crate::tui::tests::{detail_mut, state_with_gists, state_with_two_gists};
+
+    fn detail_ref(state: &AppState) -> &DetailState {
+        state.detail().expect("expected Screen::GistDetail")
+    }
+
+    fn state_with_many_files(n: usize) -> AppState {
+        let mut state = initial_state();
+        state.gists = (0..n)
+            .map(|i| GistFile {
+                gist_id: "g1".into(),
+                description: "demo".into(),
+                filename: format!("f{i}.txt"),
+                public: false,
+                updated_at: "2026-06-10T00:00:00Z".into(),
+                created_at: "2026-06-01T00:00:00Z".into(),
+                owner_login: String::new(),
+                fork_of_id: None,
+
+                raw_url: None,
+
+                content_type: None,
+
+                node_id: None,
+            })
+            .collect();
+        state
+    }
+
+    #[test]
+    fn detail_focus_and_cursor_default_to_files_and_zero() {
+        let mut state = initial_state();
+        state.screen = Screen::GistDetail(Box::default());
+        assert_eq!(detail_ref(&state).focus, DetailFocus::Files);
+        assert_eq!(detail_ref(&state).file_cursor, 0);
+    }
+
+    #[test]
+    fn detail_tab_toggles_focus() {
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        assert_eq!(detail_ref(&state).focus, DetailFocus::Files);
+        let outcome = state.handle_key(KeyCode::Tab);
+        assert!(matches!(outcome, KeyOutcome::FetchComments { .. }));
+        assert_eq!(detail_ref(&state).focus, DetailFocus::Comments);
+        let outcome = state.handle_key(KeyCode::Tab);
+        assert!(matches!(outcome, KeyOutcome::None));
+        assert_eq!(detail_ref(&state).focus, DetailFocus::Files);
+    }
+
+    #[test]
+    fn detail_tab_to_comments_skips_fetch_when_already_loaded() {
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        detail_mut(&mut state).comments = Some(Vec::new());
+        let outcome = state.handle_key(KeyCode::Tab);
+        assert!(matches!(outcome, KeyOutcome::None));
+        assert_eq!(detail_ref(&state).focus, DetailFocus::Comments);
+    }
+
+    #[test]
+    fn detail_files_focus_arrows_move_cursor_and_clamp() {
+        let mut state = state_with_gists(); // g1 has 2 files: a.txt, b.txt
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        detail_mut(&mut state).focus = DetailFocus::Files;
+
+        state.handle_key(KeyCode::Up); // already at 0, clamps
+        assert_eq!(detail_ref(&state).file_cursor, 0);
+        state.handle_key(KeyCode::Down);
+        assert_eq!(detail_ref(&state).file_cursor, 1);
+        state.handle_key(KeyCode::Down); // only 2 files, clamps at index 1
+        assert_eq!(detail_ref(&state).file_cursor, 1);
+        state.handle_key(KeyCode::PageUp); // jumps to 0
+        assert_eq!(detail_ref(&state).file_cursor, 0);
+        state.handle_key(KeyCode::PageDown); // +10 clamps to last (1)
+        assert_eq!(detail_ref(&state).file_cursor, 1);
+        // Comment scroll is untouched while files-focused.
+        assert_eq!(detail_ref(&state).scroll, 0);
+    }
+
+    #[test]
+    fn detail_comments_focus_arrows_still_scroll_comments() {
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).focus = DetailFocus::Comments;
+        state.handle_key(KeyCode::Down);
+        assert_eq!(detail_ref(&state).scroll, 1);
+        assert_eq!(detail_ref(&state).file_cursor, 0); // cursor untouched
+    }
+
+    #[test]
+    fn detail_enter_previews_cursor_file_including_tenth() {
+        let mut state = state_with_many_files(12);
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        detail_mut(&mut state).focus = DetailFocus::Files;
+        detail_mut(&mut state).file_cursor = 9; // the 10th file — unreachable via 1-9
+        let outcome = state.handle_key(KeyCode::Enter);
+        assert!(matches!(
+            outcome,
+            KeyOutcome::PreviewContent {
+                file: ref f,
+                ..
+            } if f.gist_id == "g1" && f.filename == "f9.txt"
+        ));
+        assert!(state
+            .pending_return
+            .as_ref()
+            .is_some_and(Screen::is_gist_detail));
+    }
+
+    #[test]
+    fn detail_enter_in_comments_focus_is_noop() {
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        detail_mut(&mut state).focus = DetailFocus::Comments;
+        let outcome = state.handle_key(KeyCode::Enter);
+        assert!(matches!(outcome, KeyOutcome::None));
+    }
+
+    #[test]
+    fn detail_scroll_saturates_at_zero() {
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).focus = DetailFocus::Comments;
+        detail_mut(&mut state).scroll = 0;
+        state.handle_key(KeyCode::Up);
+        assert_eq!(detail_ref(&state).scroll, 0);
+        state.handle_key(KeyCode::Down);
+        assert_eq!(detail_ref(&state).scroll, 1);
+    }
+
+    #[test]
+    fn detail_c_triggers_compaction_and_records_origin() {
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        let outcome = state.handle_key(KeyCode::Char('c'));
+        assert!(matches!(outcome, KeyOutcome::CompactGist { .. }));
+        assert!(state
+            .pending_return
+            .as_ref()
+            .is_some_and(Screen::is_gist_detail));
+    }
+
+    #[test]
+    fn detail_number_key_requests_file_preview() {
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        let outcome = state.handle_key(KeyCode::Char('1'));
+        assert!(matches!(
+            outcome,
+            KeyOutcome::PreviewContent {
+                file: ref f,
+                ..
+            } if f.gist_id == "g1" && f.filename == "a.txt"
+        ));
+        assert!(state
+            .pending_return
+            .as_ref()
+            .is_some_and(Screen::is_gist_detail));
+    }
+
+    #[test]
+    fn detail_number_key_out_of_range_is_ignored() {
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        // Only two files exist; pressing 5 must do nothing (no fetch requested).
+        let outcome = state.handle_key(KeyCode::Char('5'));
+        assert!(matches!(outcome, KeyOutcome::None));
+    }
+
+    #[test]
+    fn detail_x_requests_gist_delete_confirm() {
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        let outcome = state.handle_key(KeyCode::Char('X'));
+        assert!(matches!(outcome, KeyOutcome::None));
+        assert!(state.screen.is_confirm());
+        assert!(matches!(
+            state.pending_action(),
+            Some(PendingAction::Delete { gist_id, .. }) if gist_id == "g1"
+        ));
+    }
+
+    #[test]
+    fn star_key_in_detail_returns_toggle_intent() {
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        assert!(matches!(
+            state.handle_key(KeyCode::Char('*')),
+            KeyOutcome::ToggleGistStar { .. }
+        ));
+    }
+
+    #[test]
+    fn context_gist_id_uses_detail_id_on_detail_screen() {
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        assert_eq!(state.context_gist_id().as_deref(), Some("g1"));
+    }
+
+    #[test]
+    fn detail_e_edits_description_with_prefill_and_enter_applies() {
+        let mut state = state_with_two_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("a".into());
+        state.handle_key(KeyCode::Char('e'));
+        assert!(state.editing_description);
+        // Prefilled with the current description.
+        assert_eq!(state.description_input, "My Ghostty config");
+        state.handle_key(KeyCode::Char('!'));
+        assert_eq!(state.description_input, "My Ghostty config!");
+        assert!(matches!(
+            state.handle_key(KeyCode::Enter),
+            KeyOutcome::ApplyDescription { .. }
+        ));
+    }
+
+    #[test]
+    fn detail_description_edits_mid_string_with_cursor_keys() {
+        let mut state = state_with_two_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("a".into());
+        state.handle_key(KeyCode::Char('e'));
+        assert_eq!(state.description_input, "My Ghostty config");
+        // Jump to the start, step right past "My", and insert without retyping the rest.
+        state.handle_key(KeyCode::Home);
+        state.handle_key(KeyCode::Right);
+        state.handle_key(KeyCode::Right);
+        state.handle_key(KeyCode::Char(' '));
+        state.handle_key(KeyCode::Char('o'));
+        state.handle_key(KeyCode::Char('w'));
+        state.handle_key(KeyCode::Char('n'));
+        assert_eq!(state.description_input, "My own Ghostty config");
+        // Delete removes the char at the cursor (the space before "Ghostty").
+        state.handle_key(KeyCode::Delete);
+        assert_eq!(state.description_input, "My ownGhostty config");
+    }
+
+    #[test]
+    fn detail_esc_cancels_description_edit() {
+        let mut state = state_with_two_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("a".into());
+        state.handle_key(KeyCode::Char('e'));
+        assert!(state.editing_description);
+        state.handle_key(KeyCode::Esc);
+        assert!(!state.editing_description);
+        assert!(state.description_input.is_empty());
+    }
+
+    #[test]
+    fn detail_x_stages_whole_gist_delete() {
+        let mut state = state_with_two_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("b".into());
+        assert_eq!(state.handle_key(KeyCode::Char('X')), KeyOutcome::None);
+        assert!(state.screen.is_confirm());
+        assert_eq!(
+            state.pending_action().cloned(),
+            Some(PendingAction::Delete {
+                gist_id: "b".into(),
+                label: "SSH config".into(),
+            })
+        );
+    }
+
+    #[test]
+    fn ctrl_f_pages_gist_detail_files() {
+        use crossterm::event::KeyModifiers;
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        detail_mut(&mut state).file_cursor = 0;
+        state.handle_key_with(KeyCode::Char('f'), KeyModifiers::CONTROL);
+        assert_eq!(detail_ref(&state).file_cursor, 1);
+    }
+
+    #[test]
+    fn fork_key_returns_fork_intent_for_foreign_gist_in_detail() {
+        let mut state = initial_state();
+        state.current_user_login = Some("me".into());
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("foreign".into());
+        state.starred_gists = vec![GistFile {
+            gist_id: "foreign".into(),
+            description: "x".into(),
+            filename: "a.txt".into(),
+            public: true,
+            updated_at: "x".into(),
+            created_at: "x".into(),
+            owner_login: "other".into(),
+            fork_of_id: None,
+
+            raw_url: None,
+
+            content_type: None,
+
+            node_id: None,
+        }];
+        assert!(matches!(
+            state.handle_key(KeyCode::Char('F')),
+            KeyOutcome::ForkGist { .. }
+        ));
+    }
+
+    #[test]
+    fn fork_key_blocked_for_owned_gist_in_detail() {
+        let mut state = initial_state();
+        state.current_user_login = Some("me".into());
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("mine".into());
+        state.gists = vec![GistFile {
+            gist_id: "mine".into(),
+            description: "x".into(),
+            filename: "a.txt".into(),
+            public: true,
+            updated_at: "x".into(),
+            created_at: "x".into(),
+            owner_login: "me".into(),
+            fork_of_id: None,
+
+            raw_url: None,
+
+            content_type: None,
+
+            node_id: None,
+        }];
+        assert_eq!(state.handle_key(KeyCode::Char('F')), KeyOutcome::None);
+        assert!(state.status.as_ref().unwrap().contains("already yours"));
+    }
+
+    #[test]
+    fn foreign_detail_mutate_keys_are_silent_noop() {
+        let mut state = initial_state();
+        state.current_user_login = Some("me".into());
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("foreign".into());
+        state.starred_gists = vec![GistFile {
+            gist_id: "foreign".into(),
+            description: "x".into(),
+            filename: "a.txt".into(),
+            public: true,
+            updated_at: "x".into(),
+            created_at: "x".into(),
+            owner_login: "other".into(),
+            fork_of_id: None,
+            raw_url: None,
+            content_type: None,
+            node_id: None,
+        }];
+        assert_eq!(state.handle_key(KeyCode::Char('e')), KeyOutcome::None);
+        assert_eq!(state.handle_key(KeyCode::Char('c')), KeyOutcome::None);
+        assert_eq!(state.handle_key(KeyCode::Char('X')), KeyOutcome::None);
+        assert!(state.status.is_none());
+    }
+
+    #[test]
+    fn wheel_step_gist_detail_moves_three() {
+        // GistDetail content pane: one scroll-down tick must advance detail_scroll by 3.
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        // Use Comments focus so detail_nav moves detail_scroll (not the file cursor).
+        detail_mut(&mut state).focus = DetailFocus::Comments;
+        detail_mut(&mut state).comments = Some(Vec::new());
+        assert_eq!(detail_ref(&state).scroll, 0);
+        state.handle_mouse(MouseInput::ScrollDown, &MouseLayout::default());
+        assert_eq!(detail_ref(&state).scroll, 3);
+    }
+
+    #[test]
+    fn gist_detail_file_click_selects_and_double_previews() {
+        let mut state = state_with_gists(); // g1: a.txt (0), b.txt (1)
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        detail_mut(&mut state).focus = DetailFocus::Comments; // start elsewhere to prove the focus switch
+        let hit = PaneHit {
+            rect: Rect::new(0, 0, 40, 10),
+            offset: 0,
+        };
+        let layout = MouseLayout {
+            detail_files: Some(hit),
+            ..Default::default()
+        };
+        // Click the 2nd file row -> Files focus + cursor 1, but no open yet.
+        let out = state.handle_mouse(MouseInput::Click { col: 5, row: 2 }, &layout);
+        assert_eq!(out, KeyOutcome::None);
+        assert_eq!(detail_ref(&state).focus, DetailFocus::Files);
+        assert_eq!(detail_ref(&state).file_cursor, 1);
+        // Double-click previews that file (there is no Enter for files).
+        let out = state.handle_mouse(MouseInput::DoubleClick { col: 5, row: 2 }, &layout);
+        assert!(matches!(
+            out,
+            KeyOutcome::PreviewContent {
+                file: ref f,
+                ..
+            } if f.gist_id == "g1" && f.filename == "b.txt"
+        ));
+    }
+
+    #[test]
+    fn gist_detail_tab_click_switches_focus() {
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        detail_mut(&mut state).focus = DetailFocus::Files;
+        // Header at chunks[0] y=0: content_x = 2, tabs_y = 2; " Files " (7), " Comments " (10 @ +10).
+        let layout = MouseLayout {
+            detail_tab_files: Some(Rect::new(2, 2, 7, 1)),
+            detail_tab_comments: Some(Rect::new(12, 2, 10, 1)),
+            ..Default::default()
+        };
+        // Click the Comments tab: switches focus and (comments unloaded) requests a fetch.
+        let out = state.handle_mouse(MouseInput::Click { col: 14, row: 2 }, &layout);
+        assert_eq!(detail_ref(&state).focus, DetailFocus::Comments);
+        assert!(matches!(out, KeyOutcome::FetchComments { .. }));
+        // Click the Files tab back.
+        let out = state.handle_mouse(MouseInput::Click { col: 4, row: 2 }, &layout);
+        assert_eq!(detail_ref(&state).focus, DetailFocus::Files);
+        assert_eq!(out, KeyOutcome::None);
+    }
+
+    #[test]
+    fn wheel_step_gist_detail_files_moves_one() {
+        // The file list (Files tab) steps one file per wheel tick, not 3.
+        let mut state = state_with_gists();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        detail_mut(&mut state).focus = DetailFocus::Files;
+        detail_mut(&mut state).file_cursor = 0;
+        state.handle_mouse(MouseInput::ScrollDown, &MouseLayout::default());
+        assert_eq!(detail_ref(&state).file_cursor, 1);
+    }
+
+    #[test]
+    fn comment_lines_count_matches_view_model_thread_lines() {
+        use crate::domain::GistComment;
+        use crate::tui::screens::detail::build_gist_detail_vm;
+        use crate::tui::text::comment_lines_count;
+        use crate::tui::view_model::{CommentLineVm, CommentsPaneVm};
+        let comments = vec![
+            GistComment {
+                author: "alice".into(),
+                created_at: "2026-01-01T00:00:00Z".into(),
+                body: "one line".into(),
+            },
+            GistComment {
+                author: "bob".into(),
+                created_at: "2026-01-02T00:00:00Z".into(),
+                body: "two\nlines".into(),
+            },
+        ];
+        // Each comment: 1 header + body.lines() + 1 blank.
+        // alice: 1 + 1 + 1 = 3 ; bob: 1 + 2 + 1 = 4 ; total 7.
+        assert_eq!(comment_lines_count(&comments), 7);
+
+        let mut state = initial_state();
+        state.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut state).gist_id = Some("g1".into());
+        state.gists = vec![GistFile::for_sync("g1".into(), "a.txt".into(), None)];
+        detail_mut(&mut state).comments = Some(comments);
+        detail_mut(&mut state).comments_loaded_oldest_page = 1;
+        let detail = build_gist_detail_vm(&state);
+        match detail.comments {
+            CommentsPaneVm::Thread { lines, .. } => {
+                assert_eq!(lines.len(), 7);
+                assert!(matches!(lines[0], CommentLineVm::Author { .. }));
+            }
+            other => panic!("expected Thread, got {other:?}"),
+        }
+    }
+
+    fn sample_comment(author: &str, body: &str) -> crate::domain::GistComment {
+        crate::domain::GistComment {
+            author: author.into(),
+            created_at: "2026-01-01T00:00:00Z".into(),
+            body: body.into(),
+        }
+    }
+
+    #[test]
+    fn apply_initial_comments_sets_window_and_requests_bottom_scroll() {
+        use crate::tui::InitialComments;
+        let mut s = crate::tui::initial_state();
+        detail_mut(&mut s).gist_id = Some("g1".into());
+        s.apply_initial_comments(
+            "g1",
+            Ok(InitialComments {
+                comments: vec![sample_comment("a", "x")],
+                total: 910,
+                oldest_page: 31,
+            }),
+        );
+        assert_eq!(detail_ref(&s).comments_total, Some(910));
+        assert_eq!(detail_ref(&s).comments_loaded_oldest_page, 31);
+        assert!(detail_ref(&s).comments_scroll_to_bottom);
+        assert!(s.can_load_older_comments()); // page 31 > 1
+        assert_eq!(detail_ref(&s).comments.as_ref().unwrap().len(), 1);
+    }
+
+    #[test]
+    fn apply_initial_comments_ignored_when_gist_changed() {
+        use crate::tui::InitialComments;
+        let mut s = crate::tui::initial_state();
+        detail_mut(&mut s).gist_id = Some("g2".into());
+        s.apply_initial_comments(
+            "g1",
+            Ok(InitialComments {
+                comments: vec![],
+                total: 0,
+                oldest_page: 1,
+            }),
+        );
+        assert!(detail_ref(&s).comments.is_none()); // stale response dropped
+    }
+
+    #[test]
+    fn apply_older_comments_prepends_and_compensates_scroll() {
+        use crate::tui::InitialComments;
+        let mut s = crate::tui::initial_state();
+        detail_mut(&mut s).gist_id = Some("g1".into());
+        s.apply_initial_comments(
+            "g1",
+            Ok(InitialComments {
+                comments: vec![sample_comment("newer", "n")],
+                total: 60,
+                oldest_page: 2,
+            }),
+        );
+        detail_mut(&mut s).scroll = 5;
+        // One older comment = 1 header + 1 body + 1 blank = 3 lines prepended.
+        s.apply_older_comments("g1", Ok(vec![sample_comment("older", "o")]));
+        assert_eq!(detail_ref(&s).comments_loaded_oldest_page, 1);
+        assert!(!s.can_load_older_comments()); // reached page 1
+        assert_eq!(detail_ref(&s).comments.as_ref().unwrap()[0].author, "older"); // prepended
+        assert_eq!(detail_ref(&s).scroll, 5 + 3); // viewport held in place
+        assert!(!detail_ref(&s).comments_loading_more);
+    }
+
+    #[test]
+    fn can_load_older_false_while_loading_more() {
+        use crate::tui::InitialComments;
+        let mut s = crate::tui::initial_state();
+        detail_mut(&mut s).gist_id = Some("g1".into());
+        s.apply_initial_comments(
+            "g1",
+            Ok(InitialComments {
+                comments: vec![sample_comment("a", "x")],
+                total: 90,
+                oldest_page: 3,
+            }),
+        );
+        detail_mut(&mut s).comments_loading_more = true;
+        assert!(!s.can_load_older_comments());
+    }
+
+    #[test]
+    fn m_key_loads_older_when_available() {
+        use crate::tui::InitialComments;
+        let mut s = crate::tui::initial_state();
+        s.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut s).focus = DetailFocus::Comments;
+        detail_mut(&mut s).gist_id = Some("g1".into());
+        s.apply_initial_comments(
+            "g1",
+            Ok(InitialComments {
+                comments: vec![sample_comment("a", "x")],
+                total: 90,
+                oldest_page: 3,
+            }),
+        );
+        let out = s.handle_key(KeyCode::Char('m'));
+        assert!(matches!(out, KeyOutcome::LoadOlderComments { .. }));
+    }
+
+    #[test]
+    fn m_key_noop_when_at_oldest_page() {
+        use crate::tui::InitialComments;
+        let mut s = crate::tui::initial_state();
+        s.screen = Screen::GistDetail(Box::default());
+        detail_mut(&mut s).focus = DetailFocus::Comments;
+        detail_mut(&mut s).gist_id = Some("g1".into());
+        s.apply_initial_comments(
+            "g1",
+            Ok(InitialComments {
+                comments: vec![sample_comment("a", "x")],
+                total: 10,
+                oldest_page: 1,
+            }),
+        );
+        let out = s.handle_key(KeyCode::Char('m'));
+        assert!(matches!(out, KeyOutcome::None));
+    }
+}
