+++ src/tui/screens/gists.rs
@@ -348,3 +348,351 @@ pub(crate) fn gists_palette_items(state: &AppState) -> Vec<crate::tui::palette::
         key_item("?", "Help", KeyCode::Char('?'), true),
     ]
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::tui::*;
+
+    use crate::tui::tests::{gists_mut, set_pending, state_with_gists, state_with_two_gists};
+
+    fn gists_ref(state: &AppState) -> &GistsManagerState {
+        state.gist_manager().expect("expected Screen::Gists")
+    }
+
+    #[test]
+    fn enter_on_gist_opens_detail() {
+        let mut state = state_with_gists();
+        state.screen = Screen::Gists(Box::default());
+        let outcome = state.handle_key(KeyCode::Enter);
+        assert!(matches!(outcome, KeyOutcome::OpenGistDetail { .. }));
+    }
+
+    #[test]
+    fn context_gist_id_uses_group_cursor_on_gists_screen() {
+        let mut state = state_with_gists();
+        state.screen = Screen::Gists(Box::default());
+        gists_mut(&mut state).cursor.index = 0;
+        assert_eq!(
+            state.context_gist_id(),
+            state.selected_group().map(|g| g.id)
+        );
+    }
+
+    #[test]
+    fn o_in_gist_view_opens_browser() {
+        let mut state = state_with_two_gists();
+        state.screen = Screen::Gists(Box::default());
+        assert!(matches!(
+            state.handle_key(KeyCode::Char('o')),
+            KeyOutcome::OpenBrowser { .. }
+        ));
+    }
+
+    #[test]
+    fn compact_confirm_y_executes_and_n_returns_to_gist_manager() {
+        let mut state = state_with_two_gists();
+        // CompactAnalyze parks the restore target before Confirm opens (staged pending_return).
+        state.pending_return = Some(Screen::Gists(Box::default()));
+        set_pending(
+            &mut state,
+            PendingAction::CompactGist {
+                gist_id: "a".into(),
+                label: "My Ghostty config".into(),
+                count: 3,
+            },
+        );
+        assert_eq!(
+            state.handle_key(KeyCode::Char('y')),
+            KeyOutcome::ExecuteCompactGist
+        );
+
+        // Re-open confirm for the cancel path (y does not leave Confirm until IO runs).
+        state.pending_return = Some(Screen::Gists(Box::default()));
+        set_pending(
+            &mut state,
+            PendingAction::CompactGist {
+                gist_id: "a".into(),
+                label: "My Ghostty config".into(),
+                count: 3,
+            },
+        );
+        // Cancelling drops the pending action and lands back on the parked restore target.
+        state.handle_key(KeyCode::Char('n'));
+        assert!(state.screen.is_gists());
+        assert!(state.pending_action().is_none());
+    }
+
+    #[test]
+    fn g_opens_gist_view_landing_on_the_selected_files_gist() {
+        let mut state = state_with_two_gists();
+        // Select the second gist's row in the main (file) list, then jump to the
+        // gist-level view; it should land on that same gist.
+        state.gist_index = 1;
+        assert_eq!(state.handle_key(KeyCode::Char('g')), KeyOutcome::None);
+        assert!(state.screen.is_gists());
+        assert_eq!(gists_ref(&state).cursor.index, 1);
+        assert_eq!(state.selected_group().unwrap().id, "b");
+    }
+
+    #[test]
+    fn gist_view_v_cycles_visibility_filter() {
+        // state_with_two_gists: gist "a" is public, gist "b" is secret.
+        let mut state = state_with_two_gists();
+        state.screen = Screen::Gists(Box::default());
+        assert_eq!(state.visible_gist_groups().len(), 2);
+
+        state.handle_key(KeyCode::Char('v')); // -> public
+        let vis = state.visible_gist_groups();
+        assert_eq!(vis.len(), 1);
+        assert_eq!(vis[0].id, "a");
+
+        state.handle_key(KeyCode::Char('v')); // -> secret
+        let vis = state.visible_gist_groups();
+        assert_eq!(vis.len(), 1);
+        assert_eq!(vis[0].id, "b");
+
+        state.handle_key(KeyCode::Char('v')); // -> starred (empty source)
+        assert_eq!(state.visible_gist_groups().len(), 0);
+
+        state.handle_key(KeyCode::Char('v')); // -> forked (none here)
+        assert_eq!(state.visible_gist_groups().len(), 0);
+
+        state.handle_key(KeyCode::Char('v')); // -> all
+        assert_eq!(state.visible_gist_groups().len(), 2);
+    }
+
+    #[test]
+    fn gist_view_filter_narrows_then_esc_clears() {
+        let mut state = state_with_two_gists();
+        state.screen = Screen::Gists(Box::default());
+        state.handle_key(KeyCode::Char('/'));
+        assert!(gists_ref(&state).filtering);
+        for c in "ssh".chars() {
+            state.handle_key(KeyCode::Char(c));
+        }
+        let vis = state.visible_gist_groups();
+        assert_eq!(vis.len(), 1);
+        assert_eq!(vis[0].id, "b"); // "SSH config"
+
+        state.handle_key(KeyCode::Esc);
+        assert!(!gists_ref(&state).filtering);
+        assert!(gists_ref(&state).filter_query.is_empty());
+        assert_eq!(state.visible_gist_groups().len(), 2);
+    }
+
+    #[test]
+    fn gist_view_s_cycles_sort_updated_then_created() {
+        let mut state = initial_state();
+        state.screen = Screen::Gists(Box::default());
+        state.gists = vec![
+            GistFile {
+                gist_id: "old-upd".into(),
+                description: "x".into(),
+                filename: "f".into(),
+                public: false,
+                updated_at: "2026-01-01T00:00:00Z".into(),
+                created_at: "2026-12-01T00:00:00Z".into(),
+                owner_login: String::new(),
+                fork_of_id: None,
+
+                raw_url: None,
+
+                content_type: None,
+
+                node_id: None,
+            },
+            GistFile {
+                gist_id: "new-upd".into(),
+                description: "y".into(),
+                filename: "g".into(),
+                public: false,
+                updated_at: "2026-06-01T00:00:00Z".into(),
+                created_at: "2026-02-01T00:00:00Z".into(),
+                owner_login: String::new(),
+                fork_of_id: None,
+
+                raw_url: None,
+
+                content_type: None,
+
+                node_id: None,
+            },
+        ];
+        // Default: sort by updated (newest first).
+        assert_eq!(gists_ref(&state).sort, GistGroupSort::Updated);
+        assert_eq!(state.visible_gist_groups()[0].id, "new-upd");
+        // s -> sort by created (newest created first).
+        state.handle_key(KeyCode::Char('s'));
+        assert_eq!(gists_ref(&state).sort, GistGroupSort::Created);
+        assert_eq!(state.visible_gist_groups()[0].id, "old-upd");
+    }
+
+    #[test]
+    fn gist_view_left_right_scrolls_horizontally() {
+        let mut state = state_with_two_gists();
+        state.screen = Screen::Gists(Box::default());
+        assert_eq!(gists_ref(&state).cursor.hscroll, 0);
+        state.handle_key(KeyCode::Right);
+        assert_eq!(gists_ref(&state).cursor.hscroll, 1);
+        state.handle_key(KeyCode::Left);
+        assert_eq!(gists_ref(&state).cursor.hscroll, 0);
+        // Left at the origin saturates at 0.
+        state.handle_key(KeyCode::Left);
+        assert_eq!(gists_ref(&state).cursor.hscroll, 0);
+    }
+
+    fn gists_screen_state() -> AppState {
+        let mut state = initial_state();
+        state.gists = vec![
+            GistFile {
+                gist_id: "g1".into(),
+                description: "alpha".into(),
+                filename: "a.txt".into(),
+                public: false,
+                updated_at: "2026-06-10T00:00:00Z".into(),
+                created_at: "2026-06-01T00:00:00Z".into(),
+                owner_login: String::new(),
+                fork_of_id: None,
+
+                raw_url: None,
+
+                content_type: None,
+
+                node_id: None,
+            },
+            GistFile {
+                gist_id: "g2".into(),
+                description: "beta".into(),
+                filename: "b.txt".into(),
+                public: false,
+                updated_at: "2026-06-10T00:00:00Z".into(),
+                created_at: "2026-06-01T00:00:00Z".into(),
+                owner_login: String::new(),
+                fork_of_id: None,
+
+                raw_url: None,
+
+                content_type: None,
+
+                node_id: None,
+            },
+        ];
+        state.screen = Screen::Gists(Box::default());
+        state
+    }
+
+    #[test]
+    fn gists_filter_navigates_while_typing() {
+        let mut state = gists_screen_state();
+        gists_mut(&mut state).filtering = true;
+
+        state.handle_key(KeyCode::Down);
+        assert_eq!(gists_ref(&state).cursor.index, 1);
+        assert!(gists_ref(&state).filtering);
+        state.handle_key(KeyCode::Up);
+        assert_eq!(gists_ref(&state).cursor.index, 0);
+    }
+
+    #[test]
+    fn gists_filter_empty_backspace_exits() {
+        let mut state = gists_screen_state();
+        gists_mut(&mut state).filtering = true;
+
+        state.handle_key(KeyCode::Char('a'));
+        state.handle_key(KeyCode::Backspace); // empty again, still filtering
+        assert!(gists_ref(&state).filtering);
+        state.handle_key(KeyCode::Backspace); // empty -> exit
+        assert!(!gists_ref(&state).filtering);
+    }
+
+    #[test]
+    fn gists_filter_tab_is_noop() {
+        let mut state = gists_screen_state();
+        gists_mut(&mut state).filtering = true;
+        state.handle_key(KeyCode::Char('a'));
+
+        state.handle_key(KeyCode::Tab);
+        assert!(gists_ref(&state).filtering); // still typing
+        assert_eq!(gists_ref(&state).filter_query, "a"); // unchanged
+    }
+
+    #[test]
+    fn gists_page_keys_jump_selection() {
+        let mut state = initial_state();
+        state.screen = Screen::Gists(Box::default());
+        state.gists = (0..12)
+            .map(|i| GistFile {
+                gist_id: format!("g{i}"),
+                description: format!("gist {i}"),
+                filename: "a.txt".into(),
+                public: true,
+                updated_at: "x".into(),
+                created_at: "x".into(),
+                owner_login: String::new(),
+                fork_of_id: None,
+                raw_url: None,
+                content_type: None,
+                node_id: None,
+            })
+            .collect();
+        state.handle_key(KeyCode::PageDown);
+        assert_eq!(gists_ref(&state).cursor.index, 10);
+        state.handle_key(KeyCode::PageDown);
+        assert_eq!(gists_ref(&state).cursor.index, 11);
+    }
+
+    #[test]
+    fn fork_key_ignored_on_list_and_gist_manager() {
+        let mut state = initial_state();
+        state.current_user_login = Some("me".into());
+        state.starred_gists = vec![GistFile {
+            gist_id: "foreign".into(),
+            description: "x".into(),
+            filename: "a.txt".into(),
+            public: true,
+            updated_at: "x".into(),
+            created_at: "x".into(),
+            owner_login: "other".into(),
+            fork_of_id: None,
+
+            raw_url: None,
+
+            content_type: None,
+
+            node_id: None,
+        }];
+        state.gist_type_filter = GistTypeFilter::Starred;
+        state.gist_index = 0;
+        assert_eq!(state.handle_key(KeyCode::Char('F')), KeyOutcome::None);
+
+        state.screen = Screen::Gists(Box::default());
+        gists_mut(&mut state).type_filter = GistTypeFilter::Starred;
+        gists_mut(&mut state).cursor.index = 0;
+        assert_eq!(state.handle_key(KeyCode::Char('F')), KeyOutcome::None);
+    }
+
+    #[test]
+    fn gists_click_selects_and_double_click_matches_enter() {
+        let mut state = gists_screen_state(); // 2 groups, Screen::Gists
+        gists_mut(&mut state).cursor.index = 0;
+        let hit = PaneHit {
+            rect: Rect::new(0, 0, 40, 10),
+            offset: 0,
+        };
+        let layout = MouseLayout {
+            list: Some(hit),
+            ..Default::default()
+        };
+        // Row 2 is the 2nd content row (border at row 0) -> idx 1.
+        let out = state.handle_mouse(MouseInput::Click { col: 5, row: 2 }, &layout);
+        assert_eq!(out, KeyOutcome::None);
+        assert_eq!(gists_ref(&state).cursor.index, 1);
+        // Double-click activates the same row, exactly as Enter would.
+        let mut by_key = state.clone();
+        let key_out = by_key.handle_key(KeyCode::Enter);
+        let by_mouse = state.handle_mouse(MouseInput::DoubleClick { col: 5, row: 2 }, &layout);
+        assert_eq!(by_mouse, key_out);
+        assert!(matches!(by_mouse, KeyOutcome::OpenGistDetail { .. }));
+    }
+}
